extract isScriptingDisabled to own file
